+++ examples/basic_phoenix_ecto/lib/demo_web/router.ex
@@ -20,7 +20,6 @@ defmodule DemoWeb.Router do
     get "/", PageController, :index
     resources "/posts", PostController
 
-
     live "/users", UserLive.Index, :index
     live "/users/new", UserLive.Index, :new
     live "/users/:id/edit", UserLive.Index, :edit